refactor: move code + style + lint
