+++ back-end/src/lms/views/populate.py
@@ -447,16 +447,17 @@ def create_passports() -> dict[str, Passport]:
 
 
 def create_recruitments_offices() -> dict[str, RecruitmentOffice]:
-    values = [{
-        "title": "городов Одинцово, Звенигород, Краснознаменск и "
-                 "Одинцовского городского округа",
-        "city": "Одинцово",
-        # "district": "Одинцовский",
-    }, {
-        "title": "Московский военкомат",
-        "city": "Москва",
-        # "district": "ЦАО",
-    }]
+    values = [
+        {
+            "title": "городов Одинцово, Звенигород, Краснознаменск и "
+                     "Одинцовского городского округа",
+            "city": "Одинцово",
+        },
+        {
+            "title": "Московский военкомат",
+            "city": "Москва",
+        }
+    ]
 
     offices = {}
